xversion/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.5
Cython==3.0.9
defusedxml==0.7.1
distro==1.9.0
Django==4.2.11
django-allauth==0.61.1
django-tinymce==3.7.1
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
mysqlclient==2.2.4
oauthlib==3.2.2
openai==1.13.3
pillow==10.2.0
pycparser==2.21
pydantic==2.6.3
pydantic-core==2.16.3
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.2
typing-extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
